Updated Events Routes

+++ src/routes/event.ts
@@ -51,7 +51,7 @@ try {
         })
     }
 
-    await prisma.event.create({
+    const NewEvent = await prisma.event.create({
         data:{
             title: body.title,
             description: body.description,
@@ -63,6 +63,7 @@ try {
     })
 
     return res.status(200).json({
+        id : NewEvent.id,
         message : "Event created successfuly"
     })
 
@@ -228,23 +229,45 @@ try {
 })
 
 // Event DELETE route
-eventRouter.delete("/:id", async (req,res)=>{
+eventRouter.delete("/:id",authenticate,authorizeOrganizer,  async (req,res)=>{
  
 try {
     const EventID = req.params.id
+    const UserIdFromToken = req.user?.id
 
-    await prisma.event.delete({
+    const event = await prisma.event.findUnique({
         where:{
-            id : EventID
+            id: EventID
+        },
+        select:{
+            organizerId: true
         }
     })
+
+    if(!event){
+        return res.send(404).json({
+            message : "Event not found"
+        })
+    }
+
+    if (event.organizerId !== UserIdFromToken) {
+                console.log("hello hgellosdfksdofsdofkdsfoksfok")
+        return res.status(403).json({
+            message: "Forbidden: You can only delete your own events"
+        });
+    }
+
+    await prisma.event.delete({
+            where:{
+                id: EventID
+            }
+    })    
 
     return res.status(200).json({
         message : "Event deleted"
     })
 } catch (error) {
-    console.log(error)
-    return res.status(500).json({
+    return res.status(403).json({
         message : "Something is up with our server"
     })
 }
